updated l1-bridge
