+++ node/l1-bridge/Cargo.toml
@@ -7,10 +7,12 @@ version = "0.1.0"
 crossbeam-queue      = "0.3.12"
 futures              = "0.3"
 kaspa-consensus-core = { git = "https://github.com/kaspanet/rusty-kaspa", branch = "master" }
+kaspa-hashes         = { git = "https://github.com/kaspanet/rusty-kaspa", branch = "master" }
 kaspa-notify         = { git = "https://github.com/kaspanet/rusty-kaspa", branch = "master" }
 kaspa-rpc-core       = { git = "https://github.com/kaspanet/rusty-kaspa", branch = "master" }
 kaspa-wrpc-client    = { git = "https://github.com/kaspanet/rusty-kaspa", branch = "master" }
 log                  = "0.4"
+parking_lot          = "0.12"
 thiserror            = "2"
 tokio                = { version = "1", features = ["rt", "sync"] }
 workflow-core        = "0.18.0"
