+++ node/l1-bridge/src/event.rs
@@ -1,31 +1,42 @@
-mod block_added;
-mod daa_score_changed;
-mod finality_conflict;
-mod finality_resolved;
-mod virtual_chain_changed;
+pub use kaspa_hashes::Hash;
+use kaspa_hashes::Hash as BlockHash;
+pub use kaspa_rpc_core::RpcBlock;
 
-pub use block_added::BlockAdded;
-pub use daa_score_changed::DaaScoreChanged;
-pub use finality_conflict::FinalityConflict;
-pub use finality_resolved::FinalityResolved;
-pub use kaspa_consensus_core::Hash as BlockHash;
-pub use virtual_chain_changed::VirtualChainChanged;
-
-/// L1 event types emitted by the bridge.
+/// Events emitted by the L1 bridge.
 #[derive(Clone, Debug)]
 pub enum L1Event {
     /// Connection to L1 node established.
     Connected,
     /// Connection to L1 node lost.
     Disconnected,
-    /// A new block was added to the DAG.
-    BlockAdded(BlockAdded),
-    /// Virtual chain changed (potential reorg).
-    VirtualChainChanged(VirtualChainChanged),
-    /// Finality conflict detected.
-    FinalityConflict(FinalityConflict),
-    /// Finality conflict resolved.
-    FinalityResolved(FinalityResolved),
-    /// DAA score changed.
-    DaaScoreChanged(DaaScoreChanged),
+    /// A new block was added (in order, past to present).
+    BlockAdded {
+        /// Sequential index of this block, relative to the starting point.
+        index: u64,
+        /// The block data.
+        block: Box<RpcBlock>,
+    },
+    /// Rollback to a previous state.
+    Rollback {
+        /// Index to roll back to (this index stays, later ones removed).
+        to_index: u64,
+        /// Block hash to roll back to (this block stays, later ones removed).
+        to_hash: BlockHash,
+    },
+    /// Blocks up to this index are now finalized (pruning point advanced on L1).
+    /// The scheduler can safely prune state up to and including this index.
+    Finalized {
+        /// The index up to which blocks are finalized.
+        index: u64,
+        /// The hash of the finalized block (the new pruning point).
+        hash: BlockHash,
+    },
+    /// Bridge has completed initial sync and is now streaming live.
+    Synced,
+    /// Unrecoverable error: the starting block is no longer in the chain.
+    /// The consumer must restart the bridge from a valid checkpoint.
+    SyncLost {
+        /// Descriptive message about what happened.
+        reason: String,
+    },
 }
