+++ node/l1-bridge/src/lib.rs
@@ -4,15 +4,12 @@ mod error;
 pub mod event;
 mod event_queue;
 mod state;
+mod sync;
 
 pub use bridge::L1Bridge;
-pub use config::BridgeConfig;
+pub use config::L1BridgeConfig;
 pub use error::{L1BridgeError, Result};
-pub use event::{
-    BlockAdded, BlockHash, DaaScoreChanged, FinalityConflict, FinalityResolved, L1Event,
-    VirtualChainChanged,
-};
+pub use event::{Hash as BlockHash, L1Event, RpcBlock};
 pub use event_queue::EventQueue;
 pub use kaspa_consensus_core::network::{NetworkId, NetworkType};
 pub use kaspa_wrpc_client::prelude::ConnectStrategy;
-pub use state::ConnectionState;
