+++ node/l1-bridge/src/sync.rs
@@ -4,14 +4,14 @@ use kaspa_hashes::Hash as BlockHash;
 use kaspa_rpc_core::{RpcBlock, api::rpc::RpcApi};
 use kaspa_wrpc_client::KaspaRpcClient;
 
-use crate::{EventQueue, L2BridgeError, L2Event, Result, state::BridgeState};
+use crate::{EventQueue, L1BridgeError, L1Event, Result, state::BridgeState};
 
 /// Fetches a block by hash from the L1 node.
 pub async fn fetch_block(client: &Arc<KaspaRpcClient>, hash: BlockHash) -> Result<RpcBlock> {
     client
         .get_block(hash, true)
         .await
-        .map_err(|e| L2BridgeError::RpcCall(format!("get_block failed: {}", e)))
+        .map_err(|e| L1BridgeError::RpcCall(format!("get_block failed: {}", e)))
 }
 
 /// Gets the direct parents of a block (level 0 parents in the DAG).
@@ -37,7 +37,7 @@ pub async fn perform_initial_sync(
     let dag_info = client
         .get_block_dag_info()
         .await
-        .map_err(|e| L2BridgeError::RpcCall(format!("get_block_dag_info failed: {}", e)))?;
+        .map_err(|e| L1BridgeError::RpcCall(format!("get_block_dag_info failed: {}", e)))?;
 
     // Determine the starting point for sync.
     // If no last_processed is provided, start from the pruning point.
@@ -45,7 +45,7 @@ pub async fn perform_initial_sync(
 
     // Emit Finalized for the starting point.
     // This tells the scheduler that blocks up to initial_index are finalized on L1.
-    event_queue.push(L2Event::Finalized { index: state.initial_index(), hash: start_hash });
+    event_queue.push(L1Event::Finalized { index: state.initial_index(), hash: start_hash });
 
     // Get current virtual chain state from the starting point.
     let virtual_chain = client
@@ -55,17 +55,17 @@ pub async fn perform_initial_sync(
         )
         .await
         .map_err(|e| {
-            L2BridgeError::RpcCall(format!("get_virtual_chain_from_block failed: {}", e))
+            L1BridgeError::RpcCall(format!("get_virtual_chain_from_block failed: {}", e))
         })?;
 
     let added_hashes = virtual_chain.added_chain_block_hashes;
 
     if added_hashes.is_empty() {
-        log::info!("L2 bridge: already synced, no new blocks");
+        log::info!("L1 bridge: already synced, no new blocks");
         return Ok(last_processed);
     }
 
-    log::info!("L2 bridge: syncing {} blocks from {:?}", added_hashes.len(), last_processed);
+    log::info!("L1 bridge: syncing {} blocks from {:?}", added_hashes.len(), last_processed);
 
     // Fetch and emit blocks in order with sequential indices.
     let mut last_hash = last_processed;
@@ -75,7 +75,7 @@ pub async fn perform_initial_sync(
         last_hash = Some(block.header.hash);
         // Record the hash->index mapping for finalization tracking.
         state.record_block(block.header.hash, index);
-        event_queue.push(L2Event::BlockAdded { index, block: Box::new(block) });
+        event_queue.push(L1Event::BlockAdded { index, block: Box::new(block) });
     }
 
     Ok(last_hash)
