--- node/l2-bridge/src/error.rs
@@ -1,32 +1,0 @@
-use thiserror::Error;
-
-/// Errors that can occur in the L2 bridge.
-#[derive(Error, Debug)]
-pub enum L2BridgeError {
-    #[error("RPC connection error: {0}")]
-    Connection(String),
-
-    #[error("RPC call failed: {0}")]
-    RpcCall(String),
-
-    #[error("Subscription error: {0}")]
-    Subscription(String),
-
-    #[error("Block not found: {0}")]
-    BlockNotFound(String),
-
-    #[error("Event channel closed")]
-    ChannelClosed,
-
-    #[error("Bridge not connected")]
-    NotConnected,
-
-    #[error("Bridge already started")]
-    AlreadyStarted,
-
-    #[error("Bridge shutdown")]
-    Shutdown,
-}
-
-/// Result type for L2 bridge operations.
-pub type Result<T> = std::result::Result<T, L2BridgeError>;
