--- node/l2-bridge/src/lib.rs
@@ -1,15 +1,0 @@
-mod bridge;
-mod config;
-mod error;
-pub mod event;
-mod event_queue;
-mod state;
-mod sync;
-
-pub use bridge::L2Bridge;
-pub use config::L2BridgeConfig;
-pub use error::{L2BridgeError, Result};
-pub use event::{Hash as BlockHash, L2Event, RpcBlock};
-pub use event_queue::EventQueue;
-pub use kaspa_consensus_core::network::{NetworkId, NetworkType};
-pub use kaspa_wrpc_client::prelude::ConnectStrategy;
